webpack configuration

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/User";
 import Video from "../models/Video";
 
 export const home = async(req, res) => {
@@ -11,7 +12,7 @@ export const home = async(req, res) => {
 };
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if(!video){
         return res.status(404).render("404", { pageTitle: "Video not found." });
     }
@@ -19,19 +20,27 @@ export const watch = async (req, res) => {
 };
 export const getEdit = async(req, res) => {
     const { id } = req.params;
+    const {user: {_id}} = req.session;
     const video = await Video.findById(id);
     if(!video){
         return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 }
 
 export const postEdit = async(req, res) => {
     const { id } = req.params;
+    const {user: {_id}} = req.session;
     const { title, description, hashtags } = req.body;
     const video = await Video.exists({ _id: id });
     if(!video){
         return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -46,13 +55,20 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+    const { user: {_id} } = req.session;
+    const { path: fileUrl } = req.file;
     const { title, description, hashtags } = req.body;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
+            fileUrl,
+            owner: _id,
             hashtags: Video.formatHashtags(hashtags),
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        await User.save(); 
         return res.redirect("/");
     } catch(error) {
         return res.status(400).render("upload", {pageTitle:"Upload Video", errorMessage: error._message});
@@ -61,6 +77,14 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
+    const { user: {_id} } = req.session;
+    const video = await Video.findById(id);
+    if(!video){
+        return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 }
